requirements.txt
click==7.1.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
ddtrace==0.44.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
dnspython==2.0.0; python_version >= "3.6"
email-validator==1.1.2
flask-assets==2.0
flask-login==0.5.0
flask-sqlalchemy==2.4.4; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
flask-wtf==0.14.3
flask==1.1.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
gunicorn==20.0.4
idna==2.10; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
intervaltree==3.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
itsdangerous==1.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
jinja2==2.11.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
markupsafe==1.1.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
protobuf==3.14.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pymysql==1.0.2
python-dotenv==0.15.0
six==1.15.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
sortedcontainers==2.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
sqlalchemy==1.3.20; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
tenacity==6.2.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
uwsgi==2.0.19.1
webassets==2.0
werkzeug==1.0.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
wtforms==2.3.3
requests==2.25.1
scipy==1.5.4
keras==2.1.6
Flask-Caching==1.9.0
Flask-Session==0.3.2
pandas==1.1.5
nltk==3.5
numpy==1.16.3 #Downgrade to prevent warning statements https://github.com/tensorflow/tensorflow/issues/30427
tensorflow==1.7.0